requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.45.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pytest>=8.3.0


